+++ projects/frontend/es-core/src/hardware/Board.cpp
@@ -182,7 +182,6 @@ ICrtInterface& Board::GetCrtBoard()
 
 bool Board::CanHaveCRTBoard()
 {
-  return true;
   switch(GetBoardType())
   {
     case BoardType::Pi02: